requirements.txt
requests

django
djangorestframework
pymysql
django-cors-headers
djangorestframework-simplejwt
